Feel free to contact me, If you have any doubts
We know that machine learning is the rage these days. But the machine learning technique that shines the most brightly is deep learning. Deep learning is all about how a computer program can learn through observation and make decisions based on its experience. Deep learning methods are useful for computer vision, natural language processing, speech recognition and processing, and so much more.

The best way to learn something is with a hands-on approach is by practicing and that improve your deep learning knowledge and skills. These projects are developed with explanation and with complete code. So let’s not waste any more time and jump right into it.